modificando formulas

+++ app/src/main/java/com/example/myapplication/data/ble/FrequencyBLEReceiveManager.kt
@@ -74,7 +74,7 @@ class FrequencyBLEReceiveManager @Inject constructor(
                     this@FrequencyBLEReceiveManager.gatt = gatt
                 } else if (newState == BluetoothProfile.STATE_DISCONNECTED) {
                     coroutineScope.launch {
-                        data.emit(Resource.Success(data = FrequencyResult(0, 0, 0, ConnectionState.Disconnected)))
+                        data.emit(Resource.Success(data = FrequencyResult( 0, 0, 0, ConnectionState.Disconnected)))
                     }
                     gatt.close()
                 }
@@ -129,12 +129,12 @@ class FrequencyBLEReceiveManager @Inject constructor(
                         //XX XX XX XX XX XX
                         //23 45 12 00 12 34
                         //val multiplicator = if(value.first().toInt()> 0) -1 else 1
-                        val frequency = value[0].toInt()
+                        val frequency = value[0].toUByte()
                         val compresion = value[1].toInt()
                         //val decompresion = value[3].toInt() + value[4].toInt() / 10f
                         val position = if(value[2].toInt()> 0) 1 else -1
                         val frequencyResult = FrequencyResult(
-                            frequency,
+                            frequency.toInt(),
                             compresion,
                             position,
                             ConnectionState.Connected
